lums/tmp/benchmarking (#18)
Several updates to fully implement and improve performance of kmeans search (via ivf_hack).
* added full support for `uint8`, configurable with preprocessor in `array_types.h`
* fixed bug(s) in kmeans query routines
* parallelized kmeans query routines
* initial support for block-based partial reads of partitioned vector array
* improved (streamlined) implementation of recall computation

With this PR we have the functionality we need to do flat and ivf performance measurements for in-memory indexes.  With the EC2 instances we are currently using, we can run with up to 100M vectors.

There is still some cruft in the files (commented-out code regions) but I thought it better to get this code merged sooner rather than later since it represents a fairly substantive milestone.