Remove unit tests that were really utility programs  (#85)
This PR removes a number of tests that were really utility programs: `include/test/gen_*.cc`, `include/test/time_*.cc` and `unit_time_open.cc`.  It also cleans up a few unit tests that weren't properly configured (e.g., because a file that was expected to be in a certain place no longer is in the current repo).  Verified that all unit tests build and pass with clang on MacOS and Linux, and with g++ on Linux.