Add Type-Erased FlatL2Index and IVFFlatIndex (#218)
This PR wraps up the `flat_l2_index` and `ivf_flat_index` classes with a type-erased interface.  It recapitulates that aspect of PR #153.

The type-erased indexes allow for vector databases to be ingested and queried.  The API is modeled after those, e.g., in FAISS and pinecone.

On ingestion, the user can specify the desired types of the index.
```
  auto a = IndexIVFFlat(std::make_optional<IndexOptions>(
      {{"feature_type", "float32"},
       {"id_type", "uint32"},
       {"px_type", "uint32"}}));
```

Vector inputs can be specified in a type oblivious fashion:
```
  auto training_set = FeatureVectorArray(ctx, siftsmall_inputs_uri);
```

The index is created with 
```
  a.train(training_set, kmeans_init::random);
  a.add(training_set);
  a.write_index(ctx, api_ivf_flat_index_uri, true);
```

Queries can similarly be read and applied without specifying the type:
```
   auto query_set = FeatureVectorArray(ctx, siftsmall_query_uri);
   auto&& [s, t] = a.query_infinite_ram(query_set, k_nn, nprobe);
```

This functionality is exposed in Python via the pybind11 wrappers in `apis/python/src/tiledb/vector_search/type_erased_module.cc``.


An index can be built in a straightforward way with the Python bindings
```
        a = vspy.IndexIVFFlat(id_type="uint32", px_type="uint32")

        training_set = vspy.FeatureVectorArray(ctx, siftsmall_inputs_uri)

        a.train(training_set, "random");   # or kmeanspp
        a.add(training_set);
```

And read and queried
```
	a = vspy.IndexIVFFlat(ctx, siftsmall_group_uri)
	query_set = vspy.FeatureVectorArray(ctx, siftsmall_query_uri)

        s, t = a.query_finite_ram(query_set, k_nn, nprobe, 1000)
```