Remove unneeded clang-format CI job (#345)
### What
We don't need this because `pre-commit` runs `clang-format`. It's also failing because of a widespread issue with GitHub CI related to the apt update of Ubuntu 22.04: https://github.com/orgs/community/discussions/120966

Which you can see in my CI job here: https://github.com/TileDB-Inc/TileDB-Vector-Search/actions/runs/8815965169/job/24200539999?pr=343

So in this PR we just remove it to fix the CI pipeline.