Use unique path for temp_data folder (#357)
Vector ingestion is using a `temp_data` group that can be used to store any temporary data that are needed during the ingestion processing. This PR adds a random string in the `temp_data` uri in order for different ingestion jobs to not have conflict with each other. 

Using one `temp_data` group for all ingestion jobs was causing errors in consecutive ingestion jobs when a previous ingestion job failed or didn't cleanup its temporary data.